activity module - move some parts from app to activity module - fix broken activity form display - make activity tables scrollable
