+++ src/app/activity/forms/activity-form/activity-form-route/activity-form-route.component.html
@@ -68,10 +68,11 @@
   <div fxLayout="row" fxLayoutGap="12px" fxLayout.lt-md="column">
     <div fxFlex="60" fxLayout="column">
       <div fxFlex fxLayout="row" fxLayoutGap="12px grid">
-        <div fxFlex="25" fxLayout="column" fxLayoutAlign="start stretch">
+        <div fxFlex="50" fxLayout="column" fxLayoutAlign="start stretch">
           <mat-form-field class="no-hint">
             <mat-label>Vrsta vzpona</mat-label>
             <mat-select formControlName="ascentType">
+              <mat-select-trigger>{{ ascentTypeTrigger }}</mat-select-trigger>
               <ng-container *ngFor="let type of nonTopRopeAscentTypes">
                 <mat-option
                   *ngIf="
@@ -112,15 +113,15 @@
           </mat-form-field>
         </div>
 
-        <mat-form-field class="no-hint" fxFlex="25">
+        <!-- <mat-form-field class="no-hint" fxFlex="25">
           <mat-label>Lepota</mat-label>
           <mat-select formControlName="stars">
             <mat-option [value]="1">1</mat-option>
             <mat-option [value]="2">2</mat-option>
             <mat-option [value]="3">3</mat-option>
           </mat-select>
-        </mat-form-field>
-        <mat-form-field class="no-hint" fxFlex="25">
+        </mat-form-field> -->
+        <mat-form-field class="no-hint" fxFlex="50">
           <mat-label>Objava</mat-label>
           <mat-select formControlName="publish">
             <mat-option
@@ -130,13 +131,13 @@
             </mat-option>
           </mat-select>
         </mat-form-field>
-        <app-grade-select
+        <!-- <app-grade-select
           class="no-hint"
           fxFlex="25"
           [control]="route.controls.votedDifficulty"
           label="Predlog ocene"
           [gradingSystemId]="route.value.defaultGradingSystemId"
-        ></app-grade-select>
+        ></app-grade-select> -->
       </div>
       <div fxFlex="100" fxHide.lt-md></div>
       <div fxFlex fxLayout="row" fxLayoutGap="12px" *ngIf="!activity">
