+++ src/app/activity/forms/activity-form/activity-form.component.html
@@ -1,0 +1,89 @@
+<form [formGroup]="activityForm" (submit)="save()">
+  <div class="card">
+    <div
+      fxLayout="row"
+      fxLayout.lt-md="column"
+      fxLayoutAlign="space-between center"
+      fxLayoutAlign.lt-md="start start"
+      fxLayoutGap="16px"
+      fxLayoutGap.lt-md="0"
+    >
+      <h3>Plezalni dan</h3>
+      <mat-checkbox formControlName="onlyRoutes" class="only-routes-checkbox">
+        Vpiši samo preplezane smeri
+      </mat-checkbox>
+    </div>
+
+    <div
+      fxLayout="column"
+      fxLayoutGap="16px"
+      class="px-16"
+      *ngIf="!activityForm.value.onlyRoutes"
+    >
+      <div fxLayout="row" fxLayout.lt-md="column" fxLayoutGap="16px">
+        <mat-form-field class="no-hint" fxFlex="50">
+          <mat-label>Datum</mat-label>
+          <input matInput formControlName="date" [matDatepicker]="picker" />
+          <mat-datepicker-toggle
+            matSuffix
+            [for]="picker"
+          ></mat-datepicker-toggle>
+          <mat-datepicker #picker></mat-datepicker>
+        </mat-form-field>
+        <mat-form-field class="no-hint" fxFlex="50">
+          <mat-label>Soplezalci</mat-label>
+          <input matInput type="text" formControlName="partners" />
+        </mat-form-field>
+      </div>
+      <div fxLayout="row">
+        <mat-form-field fxFlex="100">
+          <mat-label>Dodatni zaznamki</mat-label>
+          <textarea matInput formControlName="notes"></textarea>
+        </mat-form-field>
+      </div>
+    </div>
+  </div>
+
+  <div class="card mt-16">
+    <h3>Preplezane smeri</h3>
+
+    <div class="px-16">
+      <div
+        *ngFor="
+          let route of routes.controls;
+          let first = first;
+          let last = last;
+          let index = index
+        "
+        class="route"
+      >
+        <app-activity-form-route
+          [route]="route"
+          [crag]="crag"
+          [activity]="!activityForm.value.onlyRoutes"
+          [first]="first"
+          [last]="last"
+          (move)="moveRoute(index, $event)"
+        ></app-activity-form-route>
+      </div>
+    </div>
+
+    <div fxLayout="row" fxLayoutAlign="end center" class="px-16 py-16">
+      <div *ngIf="!loading">
+        <button
+          type="button"
+          mat-button
+          [routerLink]="['/plezalisca', crag.country.slug, crag.slug]"
+        >
+          Prekliči
+        </button>
+        <button mat-button color="primary" [disabled]="!activityForm.valid">
+          SHRANI
+        </button>
+      </div>
+      <span *ngIf="loading" class="loading">
+        <mat-spinner diameter="30"></mat-spinner>
+      </span>
+    </div>
+  </div>
+</form>
