+++ src/app/activity/forms/activity-form/activity-form.component.scss
@@ -1,8 +1,7 @@
-@import "../../../styles/colors.scss";
+@import "../../../../styles/colors.scss";
 
-form {
-  max-width: 100%;
-  width: 800px;
+.only-routes-checkbox {
+  padding: 14px 16px;
 }
 
 .route {
