+++ src/app/activity/pages/activity-log/activity-log.component.scss
@@ -1,0 +1,4 @@
+.scrollable-table {
+  max-width: 100%;
+  overflow-x: auto;
+}
