+++ src/app/activity/pages/activity-routes/activity-routes.component.scss
@@ -1,0 +1,4 @@
+.scrollable-table {
+  max-width: 100%;
+  overflow-x: auto;
+}
