+++ src/app/activity/pages/activity-statistics/activity-statistics.component.html
@@ -1,5 +1,5 @@
 <app-activity-header active="statistika"></app-activity-header>
 
-<div class="container">
-  <p>activity-statistics works!</p>
+<div class="container mt-16">
+  <p><i>Statistika vzponov je še v izdelavi</i></p>
 </div>
