+++ src/app/activity/partials/activity-row/activity-row.component.ts
@@ -1,9 +1,9 @@
 import { Component, Input, OnInit } from '@angular/core';
 import { Subject } from 'rxjs';
 import { Activity } from 'src/generated/graphql';
-import { RowAction } from '../../../../activity/activity-log/activity-log.component';
 
-import { ACTIVITY_TYPES } from '../../../../common/activity.constants';
+import { ACTIVITY_TYPES } from '../../../common/activity.constants';
+import { RowAction } from '../../pages/activity-log/activity-log.component';
 
 @Component({
   selector: '[app-activity-row]',
