+++ src/app/app.module.ts
@@ -45,8 +45,6 @@ import { CragRoutesComponent } from './pages/crag/crag-routes/crag-routes.compon
 import { CragInfoComponent } from './pages/crag/crag-info/crag-info.component';
 import { CragCommentsComponent } from './pages/crag/crag-comments/crag-comments.component';
 import { MomentModule } from 'ngx-moment';
-import { ActivityFormComponent } from './forms/activity-form/activity-form.component';
-import { ActivityFormRouteComponent } from './forms/activity-form/activity-form-route/activity-form-route.component';
 import { MAT_DATE_FORMATS, MAT_DATE_LOCALE } from '@angular/material/core';
 import { MatMomentDateModule } from '@angular/material-moment-adapter';
 import { CragGalleryComponent } from './pages/crag/crag-gallery/crag-gallery.component';
@@ -59,8 +57,6 @@ import { CragImageComponent } from './pages/crag/crag-image/crag-image.component
 import { ClubsComponent } from './pages/clubs/clubs.component';
 import { MatAutocompleteModule } from '@angular/material/autocomplete';
 import { ClubMemberFormComponent } from './forms/club-member-form/club-member-form.component';
-import { AscentTypeComponent } from './common/ascent-type/ascent-type.component';
-import { AscentPublishOptionComponent } from './common/ascent-publish-option/ascent-publish-option.component';
 import { MaterialElevationDirective } from './shared/directives/material-elevation.directive';
 import { MatListModule } from '@angular/material/list';
 import { ClubComponent } from './pages/club/club.component';
@@ -106,8 +102,6 @@ const formFieldAppearance: MatFormFieldDefaultOptions = {
     CragRoutesComponent,
     CragInfoComponent,
     CragCommentsComponent,
-    ActivityFormComponent,
-    ActivityFormRouteComponent,
     CragGalleryComponent,
     RouteComponent,
     RouteInfoComponent,
@@ -117,8 +111,6 @@ const formFieldAppearance: MatFormFieldDefaultOptions = {
     CragImageComponent,
     ClubsComponent,
     ClubMemberFormComponent,
-    AscentTypeComponent,
-    AscentPublishOptionComponent,
     MaterialElevationDirective,
     ClubComponent,
     ClubMembersComponent,
