--- src/app/forms/activity-form/activity-form.component.html
@@ -1,74 +1,0 @@
-<form [formGroup]="activityForm" (submit)="save()" fxFlex fxLayout="column">
-  <div>
-    <div fxLayout="column" fxLayoutGap="12px">
-      <div fxLayout="row" fxLayoutAlign="end">
-        <mat-checkbox formControlName="onlyRoutes">
-          Vpiši samo preplezane smeri
-        </mat-checkbox>
-      </div>
-      <ng-container *ngIf="!activityForm.value.onlyRoutes">
-        <h2>Plezalni dan</h2>
-        <div fxLayout="row" fxLayoutGap="12px">
-          <mat-form-field class="no-hint" fxFlex="50">
-            <mat-label>Datum</mat-label>
-            <input matInput formControlName="date" [matDatepicker]="picker" />
-            <mat-datepicker-toggle
-              matSuffix
-              [for]="picker"
-            ></mat-datepicker-toggle>
-            <mat-datepicker #picker></mat-datepicker>
-          </mat-form-field>
-          <mat-form-field class="no-hint" fxFlex="50">
-            <mat-label>Soplezalci</mat-label>
-            <input matInput type="text" formControlName="partners" />
-          </mat-form-field>
-        </div>
-        <div fxLayout="row">
-          <mat-form-field fxFlex="100">
-            <mat-label>Dodatni zaznamki</mat-label>
-            <textarea matInput formControlName="notes"></textarea>
-          </mat-form-field>
-        </div>
-      </ng-container>
-    </div>
-
-    <h2>Preplezane smeri</h2>
-    <div>
-      <div
-        *ngFor="
-          let route of routes.controls;
-          let first = first;
-          let last = last;
-          let index = index
-        "
-        class="route"
-      >
-        <app-activity-form-route
-          [route]="route"
-          [crag]="crag"
-          [activity]="!activityForm.value.onlyRoutes"
-          [first]="first"
-          [last]="last"
-          (move)="moveRoute(index, $event)"
-        ></app-activity-form-route>
-      </div>
-    </div>
-  </div>
-  <div fxLayout="row" fxFlexAlign="end">
-    <div *ngIf="!loading">
-      <button
-        type="button"
-        mat-button
-        [routerLink]="['/plezalisce', crag.slug]"
-      >
-        Prekliči
-      </button>
-      <button mat-flat-button color="primary" [disabled]="!activityForm.valid">
-        SHRANI
-      </button>
-    </div>
-    <span *ngIf="loading" class="loading">
-      <mat-spinner diameter="30"></mat-spinner>
-    </span>
-  </div>
-</form>
