+++ src/app/shared/components/ascent-publish-option/ascent-publish-option.component.scss
