+++ src/app/shared/components/ascent-type/ascent-type.component.scss
@@ -1,4 +1,4 @@
-@import "../../../styles/colors.scss";
+@import "../../../../styles/colors.scss";
 
 .color-green {
   color: $text-green;
