+++ src/styles/global.scss
@@ -166,7 +166,7 @@ b {
 .card-table {
   width: 100%;
   .row {
-    height: 42px;
+    height: 52px;
     td {
       vertical-align: middle;
       padding: 0 12px;
